move last test from test/ into package dir
