+++ pkg/k8s/pod_cache_test.go
@@ -11,7 +11,7 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-package kiam
+package k8s
 
 import (
 	"context"
@@ -19,7 +19,6 @@ import (
 	"testing"
 	"time"
 
-	"github.com/uswitch/kiam/pkg/k8s"
 	"github.com/uswitch/kiam/pkg/testutil"
 	kt "k8s.io/client-go/tools/cache/testing"
 )
@@ -31,7 +30,7 @@ func TestFindsRunningPod(t *testing.T) {
 	defer cancel()
 
 	source := kt.NewFakeControllerSource()
-	c := k8s.NewPodCache(source, time.Second, bufferSize)
+	c := NewPodCache(source, time.Second, bufferSize)
 	source.Add(testutil.NewPodWithRole("ns", "name", "192.168.0.1", "Failed", "failed_role"))
 	source.Add(testutil.NewPodWithRole("ns", "name", "192.168.0.1", "Running", "running_role"))
 	c.Run(ctx)
@@ -50,7 +49,7 @@ func TestFindRoleActive(t *testing.T) {
 	defer cancel()
 
 	source := kt.NewFakeControllerSource()
-	c := k8s.NewPodCache(source, time.Second, bufferSize)
+	c := NewPodCache(source, time.Second, bufferSize)
 	source.Add(testutil.NewPodWithRole("ns", "name", "192.168.0.1", "Failed", "failed_role"))
 	source.Modify(testutil.NewPodWithRole("ns", "name", "192.168.0.1", "Failed", "running_role"))
 	source.Modify(testutil.NewPodWithRole("ns", "name", "192.168.0.1", "Running", "running_role"))
@@ -74,7 +73,7 @@ func BenchmarkFindPodsByIP(b *testing.B) {
 	defer cancel()
 
 	source := kt.NewFakeControllerSource()
-	c := k8s.NewPodCache(source, time.Second, bufferSize)
+	c := NewPodCache(source, time.Second, bufferSize)
 	for i := 0; i < 1000; i++ {
 		source.Add(testutil.NewPodWithRole("ns", fmt.Sprintf("name-%d", i), fmt.Sprintf("ip-%d", i), "Running", "foo_role"))
 	}
@@ -101,7 +100,7 @@ func BenchmarkIsActiveRole(b *testing.B) {
 		role := i % 100
 		source.Add(testutil.NewPodWithRole("ns", fmt.Sprintf("name-%d", i), fmt.Sprintf("ip-%d", i), "Running", fmt.Sprintf("role-%d", role)))
 	}
-	c := k8s.NewPodCache(source, time.Second, bufferSize)
+	c := NewPodCache(source, time.Second, bufferSize)
 	c.Run(ctx)
 
 	b.StartTimer()
