+++ pkg/server/policy_test.go
@@ -19,12 +19,13 @@ import (
 
 	"github.com/uswitch/kiam/pkg/aws/sts"
 	"github.com/uswitch/kiam/pkg/k8s"
+	kt "github.com/uswitch/kiam/pkg/k8s/testing"
 	"github.com/uswitch/kiam/pkg/testutil"
 )
 
 func TestRequestedRolePolicy(t *testing.T) {
 	p := testutil.NewPodWithRole("namespace", "name", "192.168.0.1", testutil.PhaseRunning, "myrole")
-	f := testutil.NewStubFinder(p)
+	f := kt.NewStubFinder(p)
 
 	arnResolver := sts.DefaultResolver("arn:aws:iam::123456789012:role/")
 	policy := NewRequestingAnnotatedRolePolicy(f, arnResolver)
@@ -61,7 +62,7 @@ func TestRequestedRolePolicy(t *testing.T) {
 func TestRequestedRolePolicyWithSlash(t *testing.T) {
 	arnResolver := sts.DefaultResolver("arn:aws:iam::123456789012:role/")
 	p := testutil.NewPodWithRole("namespace", "name", "192.168.0.1", testutil.PhaseRunning, "/myrole")
-	f := testutil.NewStubFinder(p)
+	f := kt.NewStubFinder(p)
 
 	policy := NewRequestingAnnotatedRolePolicy(f, arnResolver)
 	decision, err := policy.IsAllowedAssumeRole(context.Background(), "myrole", "192.168.0.1")
@@ -96,7 +97,7 @@ func TestRequestedRolePolicyWithSlash(t *testing.T) {
 
 func TestErrorWhenPodNotFound(t *testing.T) {
 	arnResolver := sts.DefaultResolver("arn:aws:iam::123456789012:role/")
-	f := testutil.NewStubFinder(nil)
+	f := kt.NewStubFinder(nil)
 	policy := NewRequestingAnnotatedRolePolicy(f, arnResolver)
 
 	_, err := policy.IsAllowedAssumeRole(context.Background(), "myrole", "192.168.0.1")
@@ -121,9 +122,9 @@ func TestErrorWhenPodNotFound(t *testing.T) {
 
 func TestNamespacePolicy(t *testing.T) {
 	n := testutil.NewNamespace("red", "^red.*$|^.red.*$")
-	nf := testutil.NewNamespaceFinder(n)
+	nf := kt.NewNamespaceFinder(n)
 	p := testutil.NewPodWithRole("red", "foo", "192.168.0.1", testutil.PhaseRunning, "red_role")
-	pf := testutil.NewStubFinder(p)
+	pf := kt.NewStubFinder(p)
 
 	policy := NewNamespacePermittedRoleNamePolicy(nf, pf)
 	decision, err := policy.IsAllowedAssumeRole(context.Background(), "red_role", "192.168.0.1")
@@ -158,9 +159,9 @@ func TestNamespacePolicy(t *testing.T) {
 
 func TestNamespacePolicyWithSlash(t *testing.T) {
 	n := testutil.NewNamespace("red", "^red.*$|^.red.*$")
-	nf := testutil.NewNamespaceFinder(n)
+	nf := kt.NewNamespaceFinder(n)
 	p := testutil.NewPodWithRole("red", "foo", "192.168.0.1", testutil.PhaseRunning, "/red_role")
-	pf := testutil.NewStubFinder(p)
+	pf := kt.NewStubFinder(p)
 
 	policy := NewNamespacePermittedRoleNamePolicy(nf, pf)
 	decision, err := policy.IsAllowedAssumeRole(context.Background(), "red_role", "192.168.0.1")
@@ -195,9 +196,9 @@ func TestNamespacePolicyWithSlash(t *testing.T) {
 
 func TestNotAllowedWithoutNamespaceAnnotation(t *testing.T) {
 	n := testutil.NewNamespace("red", "")
-	nf := testutil.NewNamespaceFinder(n)
+	nf := kt.NewNamespaceFinder(n)
 	p := testutil.NewPodWithRole("red", "foo", "192.168.0.1", testutil.PhaseRunning, "red_role")
-	pf := testutil.NewStubFinder(p)
+	pf := kt.NewStubFinder(p)
 
 	policy := NewNamespacePermittedRoleNamePolicy(nf, pf)
 	decision, _ := policy.IsAllowedAssumeRole(context.Background(), "red_role", "192.168.0.1")
@@ -216,9 +217,9 @@ func TestNotAllowedWithoutNamespaceAnnotation(t *testing.T) {
 
 func TestNotAllowedWithoutNamespaceAnnotationWithSlash(t *testing.T) {
 	n := testutil.NewNamespace("red", "")
-	nf := testutil.NewNamespaceFinder(n)
+	nf := kt.NewNamespaceFinder(n)
 	p := testutil.NewPodWithRole("red", "foo", "192.168.0.1", testutil.PhaseRunning, "/red_role")
-	pf := testutil.NewStubFinder(p)
+	pf := kt.NewStubFinder(p)
 
 	policy := NewNamespacePermittedRoleNamePolicy(nf, pf)
 	decision, _ := policy.IsAllowedAssumeRole(context.Background(), "red_role", "192.168.0.1")
