+++ pkg/testutil/kubernetes.go
@@ -17,7 +17,6 @@ import (
 	"fmt"
 	"time"
 
-	"github.com/uswitch/kiam/pkg/k8s"
 	"k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
@@ -33,7 +32,7 @@ func NewNamespace(name, roleRegexp string) *v1.Namespace {
 			Name: name,
 		},
 	}
-	n.Annotations = map[string]string{k8s.AnnotationPermittedKey: roleRegexp}
+	n.Annotations = map[string]string{"iam.amazonaws.com/permitted": roleRegexp}
 	return n
 }
 
@@ -57,6 +56,6 @@ func NewPod(namespace, name, ip, phase string) *v1.Pod {
 
 func NewPodWithRole(namespace, name, ip, phase, role string) *v1.Pod {
 	pod := NewPod(namespace, name, ip, phase)
-	pod.ObjectMeta.Annotations = map[string]string{k8s.AnnotationIAMRoleKey: role}
+	pod.ObjectMeta.Annotations = map[string]string{"iam.amazonaws.com/role": role}
 	return pod
 }
